# Faster calc cost matrix