- `build_C_matrix` processing 4 times faster in demo.py.
  - 146.673 sec -> 37.823 sec (in my environment)
- total processing 17% faster in demo.py.
  - 698.258 sec -> 577.944 sec (in my environment)